Removed unuseful 'connect'
Searching for an error that returned "null" from the Database. Finally it was that it is ONLY needed to do the "createConnection". The connect is not necessary to use the Databse. Furthermore, after a "connect" it has to be done an "end" because this is preventing any other future request to connect again. This is bocking the acces to the Database and limits the number of uses to only the first.